feat: Add PSR-4 autoloading support
This commit adds support for PSR-4 autoloading. The `addPsr4` method has been implemented to register a set of PSR-4 directories for a given namespace. The directories can be appended or prepended to the ones previously set for the namespace.

The `prefixDirsPsr4` property stores the registered PSR-4 prefixes and their corresponding directories, while the `fallbackDirsPsr4` property stores fallback directories that will be used if no matching prefix is found.

Additionally, new files have been added:
- composer/autoload.php
- vendor/autoload.php
- vendor/composer/autoload_classmap.php
- vendor/composer/autoload_namespaces.php
- vendor/composer/autoload_psr4.php
- vendor/composer/autoload_real.php
- vendor/composer/autoload_static.php
- vendor/composer/ClassLoader.php