refactor: Update return type and access of GitHub data
Refactored the `get_github_data()` method in the `WP_GitHub_Updater` class to update the return type from array to object. Also removed the second parameter of `json_decode()` function call to return the decoded JSON as an object instead of an array. Updated the `get_date()` and `get_description()` methods to access properties of `$data` as objects instead of arrays.